pyproject.toml
[project]
name = "myshop"
version = "0.1.0"
description = ""
authors = [
    {name = "Abzaydin",email = "abzaydin.1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "celery (>=5.3.0,<6.0.0)",
    "django (>=6.0.0,<7.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "whitenoise (>=6.6.0,<7.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "djangorestframework (>=3.15.0,<4.0.0)",
    "psycopg2-binary (>=2.9.0,<3.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-jazzmin (>=3.0.4,<4.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
